Fix review comment.
